Old spread algorithm removed
